🐛 FIX: utils functions
